pointer/requirements.txt
appdirs==1.4.3
certifi==2017.11.5
chardet==3.0.4
coverage==4.4.1
Django==1.9
djangorestframework==3.6.2
ez-setup==0.9
idna==2.6
Markdown==2.6.8
oauthlib==2.0.6
packaging==16.8
psycopg2==2.7.1
PyJWT==1.5.3
pyparsing==2.2.0
python-openid==2.2.5
requests==2.18.4
requests-oauthlib==0.8.0
six==1.10.0
social-auth-app-django==2.0.0
social-auth-core==1.5.0
urllib3==1.22
